+++ implementation/objects_and_calls/function_evaluations.h
@@ -3,7 +3,6 @@
 #include<unordered_map>
 
 #include"tuple_call_unpack.h"
-#include"call_typing.h"
 #include"object_holding.h"
 
 namespace expr
