+++ implementation/objects_and_calls/object_holding.h
@@ -1,8 +1,8 @@
 #pragma once
 
-#include"type_asker.h"
+#include"evaluator_basics/type_asker.h"
 #include"statement.h"
-#include"type_name_mapping.h"
+#include"evaluator_basics/type_name_mapping.h"
 
 namespace expr
 {
